change some bug and feature 03

+++ src/lib/myAPI/tripUtils.ts
@@ -184,11 +184,13 @@ export async function deleateBETrip(tripID: string, IDAccount: string) {
 	//     }
 	// })
 	try {
-		await prismaMongo.checkpointNSQL.deleteMany({
-			where: {
-				IDTrip: userIDTrip?.IDTrip
-			}
-		});
+		if(tripID != ''){
+			await prismaMongo.checkpointNSQL.deleteMany({
+				where: {
+					IDTrip: userIDTrip?.IDTrip
+				}
+			});
+		}
 	} catch (error) {
 		console.log('can not delete checkpointNSQL');
 	}
@@ -208,11 +210,13 @@ export async function deleateTrip(tripID: string) {
 	}
 
 	try {
+		if(tripID != ''){
 		await prismaMongo.orgChat.deleteMany({
 			where: {
 				IDTrip: tripID
 			}
 		});
+	}
 	} catch (error) {
 		console.log('can not delete orgChat');
 	}
@@ -229,11 +233,14 @@ export async function deleateTrip(tripID: string) {
 
 	//No sql deleate
 	try {
-		await prismaMongo.checkpointNSQL.deleteMany({
-			where: {
-				IDTrip: tripID
-			}
-		});
+		if(tripID != ''){
+			await prismaMongo.checkpointNSQL.deleteMany({
+				where: {
+					IDTrip: tripID
+				}
+			});
+		}
+		
 	} catch (error) {
 		console.log('can not delete checkpointNSQL');
 	}
